+++ App.js
@@ -3,13 +3,13 @@ import {NavigationContainer} from '@react-navigation/native';
 import {createStackNavigator} from '@react-navigation/stack';
 import * as firebase from 'firebase';
 import firebaseConfig from './firebase/keys';
-import WelcomeScreen from './components/WelcomeScreen';
-import SignUp from './components/SignUp';
-import SignIn from './components/SignIn';
-import LoadingScreen from './components/LoadingScreen';
+import WelcomeScreen from './components/Welcome/WelcomeScreen';
+import SignUp from './components/Welcome/SignUp';
+import SignIn from './components/Welcome/SignIn';
+import LoadingScreen from './components/Welcome/LoadingScreen';
 
 import { LogBox } from 'react-native';
-import HomeStack from './stacks/HomeStack';
+import MyTabsBar from './stacks/tabBarStack';
 
 LogBox.ignoreLogs(['Setting a timer for a long period of time'])
 
@@ -28,7 +28,7 @@ export default function App() {
       <Stack.Screen name='Welcome' component={WelcomeScreen} options={{ headerShown: false }}/>
       <Stack.Screen name='Sign Up' component={SignUp}/>
       <Stack.Screen name='Sign In' component={SignIn}/>
-      <Stack.Screen name={'Home'} component={HomeStack} options={{ headerShown: false }} />
+      <Stack.Screen name={'Home'} component={MyTabsBar} options={{ headerShown: false }} />
       </Stack.Navigator>
     </NavigationContainer>
   );