+++ components/AppScreens/Tuner.js
@@ -1,0 +1,53 @@
+import * as React from 'react';
+import { View, Text, StyleSheet, Alert} from 'react-native';
+
+
+
+export default function TunerScreen({ navigation }) {
+ 
+  return (
+    <View style={styles.container}>
+      <Text style={styles.titleText}>Tuner</Text>
+    </View>
+  )
+}
+
+const styles = StyleSheet.create({
+  button: {
+    width: 150,
+    padding: 5,
+    backgroundColor: '#ff9999',
+    borderWidth: 2,
+    borderColor: 'white',
+    borderRadius: 15,
+    alignSelf: 'center',
+  },
+  buttonText: {
+    fontSize:20,
+    color: 'white',
+    fontWeight: 'bold',
+    textAlign: 'center',
+  },
+  container: {
+    height: '100%',
+    width: '100%',
+    backgroundColor: '#fff',
+    alignItems: 'center',
+    justifyContent: 'center',
+  },
+  text: {
+    textAlign: 'center',
+    fontSize: 20,
+    fontStyle: 'italic',
+    marginTop: '2%',
+    marginBottom: '10%',
+    fontWeight: 'bold',
+    color: 'black',
+  },
+  titleText: {
+    textAlign: 'center',
+    fontSize: 30,
+    fontWeight: 'bold',
+    color: '#2E6194',
+  },
+});