+++ changelog.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
  
+## [15:19] - 2021-03-8
+ 
+### Added
+ 
+- Added Tab Bar to main app screens
+- Added SoundHub and Tuner tab options
+
+### Change
+
+- Changed the placement of the upload button to the SoundHub screen from the Welcome screen
+
 ## [00:12] - 2021-03-02
  
 ### Added